+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
 - fix: `sort-pixels`, `rgb-to-grayscale` filters
 - added `ensure_grayscale` and `grayscale_required_info` convenience methods (package: `idc.api`)
 - added `ensure_binary` and `binary_required_info` convenience methods (package: `idc.api`)
+- added `--dump_pipeline` option to `idc-convert` for saving the pipeline command
 
 
 0.0.11 (2025-04-03)